Update to nav bits flag
The calibratedPhase files now contain a variable indicating whether or not navigation bits are present in the signal. They are present only (1) if the signal is a data-tone, (2) if the signal is tracked by open loop, and (3) if no attempt is made to remove the navigation bits using navigation bit streams obtained by a ground-based receiver. The new flag is "navBitsPresent".